got the logic working for reflow
